+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 0.3.1 (unreleased)
 
+### added
+
+- plot function for `MeasurementChain` [[#288]](https://github.com/BAMWelDX/weldx/pull/288)
+
 ### ASDF
 
 - remove the `additionalProperties` restriction
@@ -9,6 +13,12 @@
 - allow scalar `integer` value in `anyOf` of `time_series-1.0.0.yaml` to
   fix [#282](https://github.com/BAMWelDX/weldx/pull/282) [[#286]](https://github.com/BAMWelDX/weldx/pull/286)
 - add examples to schema files [[#274]](https://github.com/BAMWelDX/weldx/pull/274)
+
+### fixes
+
+- A warning is now emitted if a `LocalCoordinateSystem` drops a provided time during construction. This usually happens 
+  if the coordinates and orientation only contain a single data point. 
+  [[#285]](https://github.com/BAMWelDX/weldx/pull/285)
 
 ## 0.3.0 (12.03.2021)